change back to yield_now
