+++ src/lib.rs
@@ -1,3 +1,5 @@
+// #![warn(missing_docs)]
+
 pub mod activation;
 pub mod neuralnet;
